Delete keypad menu
